feat: move token price calcs and schemas to derive utils helpers
