+++ README.MD
@@ -12,7 +12,7 @@
 
 ### Uploading to sentio
 - `npx @sentio/cli login --api-key <api-key>` (create API key in sentio dashboard)
-- `yarn sentio upload`
+- `yarn sentio upload --owner derive`
 
 ### Getting v2 assetAndSubId
 Use `chisel` in terminal
@@ -21,3 +21,5 @@ Run `bytes32(uint(uint160(<V2_ASSET_ADDRESS>)) << 96) | bytes32(<SUB_ID>)`
 ## Using derive-sentio-utils
 1. `yarn build` and `yarn link` inside of `derive-sentio-utils`
 2. `rm -rf node_modules` and `yarn cache clean` and `yarn link @derivefinance/derive-sentio-utils` and `yarn install` in subrepo that uses utils
+
+If adding new graphQL schemas, make sure to export in `index.ts` and re-run `yarn build` and `yarn sentio build`
