+++ derive-sentio-utils/abis/eth/DeriveVaultToken.json
@@ -1,0 +1,21 @@
+[
+    {
+        "type": "function",
+        "name": "getSharesValue",
+        "inputs": [
+            {
+                "name": "numShares",
+                "type": "uint256",
+                "internalType": "uint256"
+            }
+        ],
+        "outputs": [
+            {
+                "name": "",
+                "type": "uint256",
+                "internalType": "uint256"
+            }
+        ],
+        "stateMutability": "view"
+    }
+]
