+++ derive-sentio-utils/src/index.ts
@@ -2,9 +2,11 @@ import { IntegratorSeason, getCurrentSeason } from "./integratorSeasons.js";
 
 import { estimateBlockNumberAtDate } from "./crosschainBlocks.js";
 
-import { VaultConfig, V2AssetConfig, VaultPoolConfig, isVaultSubaccount } from "./vaultConfig.js";
 
 export * as constants from './constants.js';
+export * as vaults from './vaults/index.js';
 export * as v2Subgraph from './v2/v2Subgraph.js';
 
-export { IntegratorSeason, getCurrentSeason, estimateBlockNumberAtDate, VaultConfig, V2AssetConfig, VaultPoolConfig, isVaultSubaccount };
+export * as schemas from './schema/store.js';
+
+export { IntegratorSeason, getCurrentSeason, estimateBlockNumberAtDate };
