+++ etherfi-integration/package.json
@@ -11,11 +11,11 @@
     "postinstall": "sentio gen"
   },
   "dependencies": {
-    "@sentio/sdk": "^2.43.2",
+    "@sentio/sdk": "^2.44.1",
     "axios": "^1.7.4"
   },
   "devDependencies": {
-    "@sentio/cli": "^2.21.2",
+    "@sentio/cli": "^2.21.4",
     "typescript": "^5.5.2"
   }
 }
